A nit in strict cvntab
The `.value` member should be same size as the pointer-to-function
(ptrdiff_t).  This ensures it does double-duty correctly on all
platforms regardless of endianness.

+++ getpar.h
@@ -35,6 +35,8 @@
 #ifndef TREK_GET_PAR_H
 #define TREK_GET_PAR_H
 
+#include <stddef.h>
+
 typedef void (*cvntab_fn)(int);
 
 /* used for getcodpar() paramater list */
@@ -45,7 +47,7 @@ struct cvntab
     union
     {
         cvntab_fn func;
-        long long value;
+        ptrdiff_t value;
     };
     int value2;
 };
